refactor(app): extract theme-switch hooks + UI out of route file (IM4, CB1)
Move theme hooks to app/hooks/useTheme.ts and ThemeSwitch to
app/components/ThemeSwitch; route retains only schema + action.
Hoist IIFE to const in ThemeSwitch render (SG12).
